Set up front end and install dependencies

+++ backend/controllers/productController.js
@@ -61,13 +61,19 @@ const updateProductController = async (req, res) => {
 
 const deleteProductController = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Invalid product id or product id not found!",
+    });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
       .status(200)
       .json({ success: true, message: "Product deleted successfully!" });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product Not Found!" });
+    res.status(500).json({ success: false, message: "Internal Server Error" });
   }
 };
